+++ traffic_ops_ort/atstccfg/toreq/client.go
@@ -51,16 +51,20 @@ type TOClient struct {
 func New(url *url.URL, user string, pass string, insecure bool, timeout time.Duration, userAgent string) (*TOClient, error) {
 	log.Infoln("URL: '" + url.String() + "' User: '" + user + "' Pass len: '" + strconv.Itoa(len(pass)) + "'")
 
-	toFQDN := url.Scheme + "://" + url.Host
-	log.Infoln("TO FQDN: '" + toFQDN + "'")
+	toURLStr := url.Scheme + "://" + url.Host
+	log.Infoln("TO URL string: '" + toURLStr + "'")
 	log.Infoln("TO URL: '" + url.String() + "'")
 
-	toClient, toIP, err := toclient.LoginWithAgent(toFQDN, user, pass, insecure, userAgent, false, timeout)
+	opts := toclient.ClientOpts{}
+	opts.Insecure = insecure
+	opts.UserAgent = userAgent
+	opts.RequestTimeout = timeout
+	toClient, inf, err := toclient.Login(toURLStr, user, pass, opts)
 	if err != nil {
-		return nil, errors.New("Logging in to Traffic Ops '" + torequtil.MaybeIPStr(toIP) + "': " + err.Error())
+		return nil, errors.New("Logging in to Traffic Ops '" + torequtil.MaybeIPStr(inf.RemoteAddr) + "': " + err.Error())
 	}
 
-	log.Infoln("toreqnew.New Logged into in to Traffic Ops '" + torequtil.MaybeIPStr(toIP) + "'")
+	log.Infoln("toreqnew.New Logged into in to Traffic Ops '" + torequtil.MaybeIPStr(inf.RemoteAddr) + "'")
 
 	latestSupported, toAddr, err := IsLatestSupported(toClient)
 	if err != nil {
@@ -69,7 +73,7 @@ func New(url *url.URL, user string, pass string, insecure bool, timeout time.Dur
 
 	client := &TOClient{C: toClient}
 	if !latestSupported {
-		log.Warnln("toreqnew.New Traffic Ops '" + torequtil.MaybeIPStr(toIP) + "' does not support the latest client, falling back ot the previous")
+		log.Warnln("toreqnew.New Traffic Ops '" + torequtil.MaybeIPStr(inf.RemoteAddr) + "' does not support the latest client, falling back ot the previous")
 
 		oldClient, err := toreqold.New(url, user, pass, insecure, timeout, userAgent)
 		if err != nil {